Code restructuring for better usability from other projects
